requirements.txt
selenium
pandas
openpyxl
pyautogui
pyperclip
streamlit
Pillow
